Finish FK and IK testing for YASKAWA but maybe not suit for other brand. Notice: must be careful about the accuracy of number in C++.

+++ KinemicsCppTester/Extension.cpp
@@ -18,7 +18,7 @@ double toDeg(double rad) {
 
 float round(float num, int index) {
 	auto multiplier = pow(10, index);
-	return round(num * multiplier) / multiplier;
+	return roundf(num * multiplier) / multiplier;
 };
 
 double round(double num, int index) {
